refactor: Move orientation-related code to separate file
